changed replicate

+++ opt/cassandra/test/distributed/org/apache/cassandra/distributed/impl/AbstractCluster.java
@@ -46,6 +46,7 @@ import org.slf4j.LoggerFactory;
 import org.apache.cassandra.db.ColumnFamilyStore;
 import org.apache.cassandra.db.ConsistencyLevel;
 import org.apache.cassandra.db.Keyspace;
+import org.apache.cassandra.distributed.api.Feature;
 import org.apache.cassandra.distributed.api.ICoordinator;
 import org.apache.cassandra.distributed.api.IInstance;
 import org.apache.cassandra.distributed.api.IInstanceConfig;
@@ -102,6 +103,7 @@ public abstract class AbstractCluster<I extends IInstance> implements ICluster, 
 
     // mutated by user-facing API
     private final MessageFilters filters;
+    private volatile Thread.UncaughtExceptionHandler previousHandler = null;
 
     protected class Wrapper extends DelegatingInvokableInstance implements IUpgradeableInstance
     {
@@ -129,7 +131,7 @@ public abstract class AbstractCluster<I extends IInstance> implements ICluster, 
 
         private IInvokableInstance newInstance(int generation)
         {
-            ClassLoader classLoader = new InstanceClassLoader(generation, version.classpath, sharedClassLoader);
+            ClassLoader classLoader = new InstanceClassLoader(generation, config.num, version.classpath, sharedClassLoader);
             return Instance.transferAdhoc((SerializableBiFunction<IInstanceConfig, ClassLoader, Instance>)Instance::new, classLoader)
                                         .apply(config.forVersion(version.major), classLoader);
         }
@@ -171,6 +173,28 @@ public abstract class AbstractCluster<I extends IInstance> implements ICluster, 
             return future;
         }
 
+        public int liveMemberCount()
+        {
+            if (!isShutdown && delegate != null)
+                return delegate().liveMemberCount();
+
+            throw new IllegalStateException("Cannot get live member count on shutdown instance");
+        }
+
+        public int nodetool(String... commandAndArgs)
+        {
+            return delegate().nodetool(commandAndArgs);
+        }
+
+        public long killAttempts()
+        {
+            IInvokableInstance local = delegate;
+            // if shutdown cleared the delegate, then no longer know how many kill attempts happened, so return -1
+            if (local == null)
+                return -1;
+            return local.killAttempts();
+        }
+
         @Override
         public void receiveMessage(IMessage message)
         {
@@ -192,6 +216,15 @@ public abstract class AbstractCluster<I extends IInstance> implements ICluster, 
                 delegate.shutdown();
                 delegate = null;
             }
+        }
+
+        public void uncaughtException(Thread thread, Throwable throwable)
+        {
+            IInvokableInstance delegate = this.delegate;
+            if (delegate != null)
+                delegate.uncaughtException(thread, throwable);
+            else
+                logger.error("uncaught exception in thread {}", thread, throwable);
         }
     }
 
@@ -250,15 +283,21 @@ public abstract class AbstractCluster<I extends IInstance> implements ICluster, 
     }
 
     public void forEach(IIsolatedExecutor.SerializableRunnable runnable) { forEach(i -> i.sync(runnable)); }
-    public void forEach(Consumer<? super I> consumer) { instances.forEach(consumer); }
-    public void parallelForEach(IIsolatedExecutor.SerializableConsumer<? super I> consumer, long timeout, TimeUnit units)
+    public void forEach(Consumer<? super I> consumer) { forEach(instances, consumer); }
+    public void forEach(List<I> instancesForOp, Consumer<? super I> consumer) { instancesForOp.forEach(consumer); }
+
+    public void parallelForEach(IIsolatedExecutor.SerializableConsumer<? super I> consumer, long timeout, TimeUnit unit)
+    {
+            parallelForEach(instances, consumer, timeout, unit);
+    }
+
+    public void parallelForEach(List<I> instances, IIsolatedExecutor.SerializableConsumer<? super I> consumer, long timeout, TimeUnit unit)
     {
         FBUtilities.waitOnFutures(instances.stream()
                                            .map(i -> i.async(consumer).apply(i))
                                            .collect(Collectors.toList()),
-                                  timeout, units);
+                                  timeout, unit);
     }
-
 
     public IMessageFilters filters() { return filters; }
     public MessageFilters.Builder verbs(Verb... verbs) { return filters.verbs(verbs); }
@@ -278,7 +317,7 @@ public abstract class AbstractCluster<I extends IInstance> implements ICluster, 
             {
                 // execute the schema change
                 coordinator(1).execute(query, ConsistencyLevel.ALL);
-                monitor.waitForAgreement();
+                monitor.waitForCompletion();
             }
         }).run();
     }
@@ -301,6 +340,66 @@ public abstract class AbstractCluster<I extends IInstance> implements ICluster, 
         }
     }
 
+    public abstract class ChangeMonitor implements AutoCloseable
+    {
+        final List<IListen.Cancel> cleanup;
+        final SimpleCondition completed;
+        private final long timeOut;
+        private final TimeUnit timeoutUnit;
+        volatile boolean changed;
+
+        public ChangeMonitor(long timeOut, TimeUnit timeoutUnit)
+        {
+            this.timeOut = timeOut;
+            this.timeoutUnit = timeoutUnit;
+            this.cleanup = new ArrayList<>(instances.size());
+            this.completed = new SimpleCondition();
+        }
+
+        protected void signal()
+        {
+            if (changed && isCompleted())
+                completed.signalAll();
+        }
+
+        @Override
+        public void close()
+        {
+            for (IListen.Cancel cancel : cleanup)
+                cancel.cancel();
+        }
+
+        public void waitForCompletion()
+        {
+            startPolling();
+            changed = true;
+            signal();
+            try
+            {
+                if (!completed.await(timeOut, timeoutUnit))
+                    throw new InterruptedException();
+            }
+            catch (InterruptedException e)
+            {
+                throw new IllegalStateException(getMonitorTimeoutMessage());
+            }
+        }
+
+        private void startPolling()
+        {
+            for (IInstance instance : instances)
+                cleanup.add(startPolling(instance));
+        }
+
+        protected abstract IListen.Cancel startPolling(IInstance instance);
+
+        protected abstract boolean isCompleted();
+
+        protected abstract String getMonitorTimeoutMessage();
+    }
+
+
+
     /**
      * Will wait for a schema change AND agreement that occurs after it is created
      * (and precedes the invocation to waitForAgreement)
@@ -311,45 +410,49 @@ public abstract class AbstractCluster<I extends IInstance> implements ICluster, 
      *
      * This could perhaps be made a little more robust, but this should more than suffice.
      */
-    public class SchemaChangeMonitor implements AutoCloseable
+    public class SchemaChangeMonitor extends ChangeMonitor
     {
-        final List<IListen.Cancel> cleanup;
-        volatile boolean schemaHasChanged;
-        final SimpleCondition agreement = new SimpleCondition();
-
         public SchemaChangeMonitor()
         {
-            this.cleanup = new ArrayList<>(instances.size());
-            for (IInstance instance : instances)
-                cleanup.add(instance.listen().schema(this::signal));
+            super(70, TimeUnit.SECONDS);
         }
 
-        private void signal()
+        protected IListen.Cancel startPolling(IInstance instance)
         {
-            if (schemaHasChanged && 1 == instances.stream().filter(i -> !i.isShutdown()).map(IInstance::schemaVersion).distinct().count())
-                agreement.signalAll();
+            return instance.listen().schema(this::signal);
         }
 
-        @Override
-        public void close()
+        protected boolean isCompleted()
         {
-            for (IListen.Cancel cancel : cleanup)
-                cancel.cancel();
+            return 1 == instances.stream().map(IInstance::schemaVersion).distinct().count();
         }
 
-        public void waitForAgreement()
+        protected String getMonitorTimeoutMessage()
         {
-            schemaHasChanged = true;
-            signal();
-            try
-            {
-                if (!agreement.await(1L, TimeUnit.MINUTES))
-                    throw new InterruptedException();
-            }
-            catch (InterruptedException e)
-            {
-                throw new IllegalStateException("Schema agreement not reached");
-            }
+            return "Schema agreement not reached";
+        }
+    }
+
+    public class AllMembersAliveMonitor extends ChangeMonitor
+    {
+        public AllMembersAliveMonitor()
+        {
+            super(60, TimeUnit.SECONDS);
+        }
+
+        protected IListen.Cancel startPolling(IInstance instance)
+        {
+            return instance.listen().liveMembers(this::signal);
+        }
+
+        protected boolean isCompleted()
+        {
+            return instances.stream().allMatch(i -> !i.config().has(Feature.GOSSIP) || i.liveMemberCount() == instances.size());
+        }
+
+        protected String getMonitorTimeoutMessage()
+        {
+            return "Live member count did not converge across all instances";
         }
     }
 
@@ -360,7 +463,40 @@ public abstract class AbstractCluster<I extends IInstance> implements ICluster, 
 
     void startup()
     {
-        parallelForEach(I::startup, 0, null);
+        previousHandler = Thread.getDefaultUncaughtExceptionHandler();
+        Thread.setDefaultUncaughtExceptionHandler(this::uncaughtExceptions);
+        try (AllMembersAliveMonitor monitor = new AllMembersAliveMonitor())
+        {
+            // Start any instances with auto_bootstrap enabled first, and in series to avoid issues
+            // with multiple nodes bootstrapping with consistent range movement enabled,
+            // and then start any instances with it disabled in parallel.
+            List<I> startSequentially = new ArrayList<>();
+            List<I> startParallel = new ArrayList<>();
+            for (I instance : instances)
+            {
+                if ((boolean) instance.config().get("auto_bootstrap"))
+                    startSequentially.add(instance);
+                else
+                    startParallel.add(instance);
+            }
+
+            forEach(startSequentially, I::startup);
+            parallelForEach(startParallel, I::startup, 0, null);
+            monitor.waitForCompletion();
+        }
+    }
+
+    private void uncaughtExceptions(Thread thread, Throwable error)
+    {
+        if (!(thread.getContextClassLoader() instanceof InstanceClassLoader))
+        {
+            Thread.UncaughtExceptionHandler handler = previousHandler;
+            if (null != handler)
+                handler.uncaughtException(thread, error);
+            return;
+        }
+        InstanceClassLoader cl = (InstanceClassLoader) thread.getContextClassLoader();
+        get(cl.getInstanceId()).uncaughtException(thread, error);
     }
 
     protected interface Factory<I extends IInstance, C extends AbstractCluster<I>>
@@ -526,6 +662,7 @@ public abstract class AbstractCluster<I extends IInstance> implements ICluster, 
             long token = Long.MIN_VALUE + 1, increment = 2 * (Long.MAX_VALUE / nodeCount);
 
             String ipPrefix = "127.0." + subnet + ".";
+            String seedIp = ipPrefix + "1";
 
             NetworkTopology networkTopology = NetworkTopology.build(ipPrefix, 7012, nodeIdTopology);
 
@@ -533,7 +670,7 @@ public abstract class AbstractCluster<I extends IInstance> implements ICluster, 
             {
                 int nodeNum = i + 1;
                 String ipAddress = ipPrefix + nodeNum;
-                InstanceConfig config = InstanceConfig.generate(i + 1, ipAddress, networkTopology, root, String.valueOf(token));
+                InstanceConfig config = InstanceConfig.generate(i + 1, ipAddress, networkTopology, root, String.valueOf(token), seedIp);
                 if (configUpdater != null)
                     configUpdater.accept(config);
                 configs.add(config);
@@ -594,7 +731,10 @@ public abstract class AbstractCluster<I extends IInstance> implements ICluster, 
         instances.clear();
         instanceMap.clear();
         // Make sure to only delete directory when threads are stopped
-        FileUtils.deleteRecursive(root);
+        if (root.exists())
+            FileUtils.deleteRecursive(root);
+        Thread.setDefaultUncaughtExceptionHandler(previousHandler);
+        previousHandler = null;
 
         //withThreadLeakCheck(futures);
     }
